+++ sapao-na-cozinha/package.json
@@ -13,6 +13,7 @@
     "@tailwindcss/vite": "^4.0.11",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.3.0",
     "tailwindcss": "^4.0.11"
   },
   "devDependencies": {
