+++ sapao-na-cozinha/src/App.jsx
@@ -1,16 +1,15 @@
 import './App.css'
+import { Routes, Route } from "react-router-dom";
 
-import PaginaCadastro from './components/PaginaCadastro';
-import PaginaLogin from './components/PaginaLogin.jsx';
-import PaginaInicial from './components/PaginaInicial';
+import PaginaCadastro from './components/pages/PaginaCadastro.jsx';
+import PaginaLogin from './components/pages/PaginaLogin.jsx';
 
 function App() {
   return (
-    <div className="flex items-center justify-center min-h-screen bg-[var(--color-quinacridone-magenta)] bg-[url('./assets/images/fundo_32021F.png')] bg-repeat bg-[size:500px]">
-      <PaginaCadastro />
-    <div className='flex items-center justify-center min-h-screen bg-[#e7e5e4]'>
-      <PaginaInicial/>
-    </div>
+      <Routes>
+        <Route path="/" element={<PaginaLogin />} />
+        <Route path="/cadastro" element={<PaginaCadastro />} />
+      </Routes>
   );
 }
