+++ sapao-na-cozinha/src/components/pages/main-page/MainPage.jsx
@@ -1,0 +1,9 @@
+import PaginaInicial from './components/pages/main-page/PaginaInicial.jsx';
+
+export default function MainPage() {
+    return (
+        <div className='flex items-center justify-center min-h-screen bg-[#e7e5e4]'>
+            <PaginaInicial />
+        </div>
+    )
+}
